pyproject.toml
[tool.poetry]
name = "django-jira"
version = "0.1.40"
description = "Intergrates Jira library into django microservices"
authors = ["Antoine Wood <antoinewood@gmail.com>"]
readme = "README.md"
include = []

[tool.poetry.dependencies]
python = "^3.8"
jira = "^3.4.1"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
djangorestframework = "^3.13.1"
Django = "^3.2.0"
Jinja2 = "^3.1.2"


[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
nox = "^2022.8.7"
pytest = "^7.1.3"
flake8 = "^5.0.4"
flake8-bandit = "^4.1.1"
mypy = "^0.971"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["src"]